refactor windows module
